EDU-904 Move resource-picker related components into one folder
