+++ src/components/resource-picker/files-grid-viewer.less
