+++ src/components/resource-picker/files-list-viewer.less
