+++ src/components/resource-picker/files-viewer.js
@@ -4,8 +4,8 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import FilesGridViewer from './files-grid-viewer.js';
 import FilesListViewer from './files-list-viewer.js';
-import { cdnObjectShape } from '../ui/default-prop-types.js';
-import { FILES_VIEWER_DISPLAY } from '../domain/constants.js';
+import { cdnObjectShape } from '../../ui/default-prop-types.js';
+import { FILES_VIEWER_DISPLAY } from '../../domain/constants.js';
 
 function FilesViewer({
   display,
