+++ src/components/resource-picker/files-viewer.less
