+++ src/components/resource-picker/resource-preview-screen.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Button } from 'antd';
 import PropTypes from 'prop-types';
-import FilePreview from '../file-preview.js';
 import { useTranslation } from 'react-i18next';
+import ResourcePreview from './resource-preview.js';
 import { ArrowLeftOutlined } from '@ant-design/icons';
 
 function PreviewScreen({ file, onBackClick, onCancelClick, onSelectClick }) {
@@ -12,7 +12,7 @@ function PreviewScreen({ file, onBackClick, onCancelClick, onSelectClick }) {
     <div className="u-resource-picker-screen">
       <h3>{t('headline')}</h3>
       <div className="u-resource-picker-screen-content u-resource-picker-screen-content-fit u-resource-picker-screen-content-scrollable">
-        <FilePreview url={file.url} size={file.size} createdOn={file.createdOn || null} updatedOn={file.updatedOn || null} />
+        <ResourcePreview url={file.url} size={file.size} createdOn={file.createdOn || null} updatedOn={file.updatedOn || null} />
       </div>
       <div className="u-resource-picker-screen-footer">
         <Button onClick={onBackClick} icon={<ArrowLeftOutlined />}>{t('common:back')}</Button>
