+++ src/components/resource-picker/resource-preview.less
