+++ src/styles/main.less
@@ -42,9 +42,7 @@
 @import '../components/pages/users.less';
 @import '../components/color-picker.less';
 @import '../components/pdf-document.less';
-@import '../components/file-preview.less';
 @import '../components/used-storage.less';
-@import '../components/files-viewer.less';
 @import '../components/tag-selector.less';
 @import '../components/abc-notation.less';
 @import '../components/pages/search.less';
@@ -73,8 +71,6 @@
 @import '../components/copyright-notice.less';
 @import '../components/default-site-logo.less';
 @import '../components/input-and-preview.less';
-@import '../components/files-grid-viewer.less';
-@import '../components/files-list-viewer.less';
 @import '../components/edit-control-panel.less';
 @import '../components/ui-language-dialog.less';
 @import '../components/admin/settings-tab.less';
@@ -95,13 +91,17 @@
 @import '../components/localization/language-icon.less';
 @import '../components/pages/complete-registration.less';
 @import '../components/pages/browser-not-supported.less';
+@import '../components/resource-picker/files-viewer.less';
 @import '../components/pages/complete-password-reset.less';
 @import '../components/media-player/media-player-track.less';
 @import '../components/admin/technical-maintenance-tab.less';
+@import '../components/resource-picker/resource-preview.less';
 @import '../components/media-player/media-volume-slider.less';
 @import '../components/resource-picker/storage-location.less';
 @import '../components/media-player/media-range-selector.less';
 @import '../components/resource-picker/resource-selector.less';
+@import '../components/resource-picker/files-grid-viewer.less';
+@import '../components/resource-picker/files-list-viewer.less';
 @import '../components/media-player/volume-presets-modal.less';
 @import '../components/media-player/media-player-controls.less';
 @import '../components/resource-picker/file-editor-screen.less';
